renamed the pages directory to screens
